+++ src/components/Explore/types.ts
@@ -1,0 +1,10 @@
+export enum Category {
+  percent_change = '% Change',
+  market_cap = 'Market Cap',
+  price = 'Price',
+  volume = 'Volume',
+}
+export enum SortDirection {
+  Increasing = 'Increasing',
+  Decreasing = 'Decreasing',
+}
